Move html email util function
