--- src/django/users/templates/registration/password_reset_email.txt
@@ -1,5 +1,0 @@
-{% load i18n %}
-{% blocktrans %}Reset password at {{ site_name }}{% endblocktrans %}:
-{% block reset_link %}
-https://{{ domain }}{% url 'auth_password_reset_confirm' uid token %}
-{% endblock %}
